LayoutTests/crypto/workers/subtle/aes-generate-export-key-jwk-expected.txt

[Worker] Test exporting an AES key using AES-CBC algorithm in workers with JWK format.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Starting worker: resources/aes-generate-export-key-jwk.js
[Worker] Generating a key...
[Worker] Exporting a key...
PASS [Worker] key.kty is 'oct'
PASS [Worker] key.key_ops is ['decrypt', 'encrypt', 'unwrapKey', 'wrapKey']
PASS [Worker] key.alg is 'A128CBC'
PASS [Worker] key.ext is true
PASS [Worker] Base64URL.parse(key.k).byteLength is 16
PASS successfullyParsed is true

TEST COMPLETE
